fix: addressed my own feedbacks
