+++ app/products/page.tsx
@@ -1,4 +1,3 @@
-"use client";
 import { ProductCard } from "@/components/product-card"
 import { products } from "@/lib/products"
