+++ package.json
@@ -12,7 +12,6 @@
     "@radix-ui/react-dropdown-menu": "^2.1.6",
     "@radix-ui/react-slot": "^1.1.2",
     "class-variance-authority": "^0.7.1",
-    "dotenv": "^16.4.5",
     "lucide-react": "^0.477.0",
     "next": "14.2.16",
     "react": "^18",
